requirements.txt
slackclient
flask==2.0.2
slackeventsapi==3.0.1
pydrive==1.3.1
treelib==1.6.1
treetojson==0.9.1
pandas
gspread_pandas